refactor(app): make app file name lowercase
