+++ src/main.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import Router from 'vue-router';
-import App from './App.vue';
+import App from './app.vue';
 import vuetify from './plugins/vuetify';
 
 Vue.use(Router);
